+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v9.2.0
+- Add and integrate `MacClientInfo` proto message. (#57)
 - Add metrics component to measure SDK performance. (#77)
 
 # v9.1.4